Change file structure and add more documentation on RTT patch
